+++ dbt/include/trino/macros/crossdb/any_value.sql
@@ -1,0 +1,3 @@
+{% macro default__any_value(expression) -%}
+    min({{ expression }})
+{%- endmacro %}