+++ dbt/include/trino/macros/crossdb/current_timestamp.sql
@@ -1,0 +1,3 @@
+{% macro trino__current_timestamp() %}
+    current_timestamp
+{% endmacro %}