+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ### Fixes
 - Fix `TrinoKerberosCredentials.trino_auth()` for kerberos authentication ([#78](https://github.com/starburstdata/dbt-trino/pull/78))
 ### Under the hood
+- Move crossdb macros from trino-dbt-utils to adapter ([#83](https://github.com/starburstdata/dbt-trino/pull/83))
 
 Contributors:
 * [@vivianhsu0214](https://github.com/vivianhsu0214) ([#78](https://github.com/starburstdata/dbt-trino/pull/78))
+* [@mdesmet](https://github.com/mdesmet) ([#78](https://github.com/starburstdata/dbt-trino/pull/83))
 
 ## dbt-trino 1.1.1 (June 20, 2022)